requirements.txt
tensorflow==2.19.0
tf-keras==2.19.0
streamlit==1.44.0
sentence-transformers==3.4.1
chromadb==0.6.3
Pillow==11.1.0
numpy==2.1.3
groq==0.20.0
